Move login components into submodules.
