+++ client/src/container/login/Login.tsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 
 import { Redirect } from 'react-router-dom';
 
-import LoginPage from '../component/LoginPage';
+import LoginPage from '../../component/login/LoginPage';
 
 
 interface LoginContainerProps {
